+++ src/components/imageSlide/ImageSlide.js
@@ -9,7 +9,7 @@ import SwiperCore, { Navigation } from "swiper";
 // 컴포넌트
 import AddButton from "./AddButton"
 
-const ImageSlide = ({setSelect, select, setImgUrl, imgUrl, setImgs, imgs, l, j, setFocus, focus}) => {
+const ImageSlide = ({editdata, setSelect, select, setImgUrl, imgUrl, setImgs, imgs, l, j, setFocus, focus}) => {
   SwiperCore.use([Navigation]);
  
  const loadImg = async (e, index) => {
@@ -46,7 +46,7 @@ const ImageSlide = ({setSelect, select, setImgUrl, imgUrl, setImgs, imgs, l, j, 
     {focus&&focus.length !== 0 ?
       <>
       <div className="writeImageContainerPerPlaceWrap"
-      style={imgUrl&&imgUrl[j]&&imgUrl[j].imgUrl.length !== 0? {display:"flex"}:{display:"none"}}
+      style={imgUrl[j]&&imgUrl[j].imgUrl&&imgUrl[j].imgUrl.length !== 0 ? {display:"flex"}:{display:"none"}}
       >
         <Swiper 
         style={{
@@ -79,7 +79,10 @@ const ImageSlide = ({setSelect, select, setImgUrl, imgUrl, setImgs, imgs, l, j, 
     :
 
     <>
-    <div className="writeImageContainerPerPlaceWrap">
+    <div className="writeImageContainerPerPlaceWrap"
+    style={imgUrl[0]&&imgUrl[0].imgUrl&&imgUrl[0].imgUrl.length !== 0 ? {display:"flex"}:{display:"none"}}
+    >
+      {/* 지도에 장소핀은 있는데 아직 클릭은 하지 않았을 때 첫번째 장소 이미지 띄워주기 */}
       <Swiper 
       style={{
         width : "100%",
@@ -93,14 +96,14 @@ const ImageSlide = ({setSelect, select, setImgUrl, imgUrl, setImgs, imgs, l, j, 
         300: {
             slidesPerView: 1
         }}}>
-      {imgUrl[j].imgUrl.map((list, i) => 
+      {imgUrl&&imgUrl[0]&&imgUrl[0].imgUrl.map((list, i) => 
       <SwiperSlide 
         style={{
           width : "343px",
           height: "256px"
         }}
       className="write_categoryslide-imagecontent" key={i}>
-         <img src={list} alt="장소이미지" style={{width:"343px"}}/>
+        <img src={list} alt="장소이미지" style={{width:"343px"}}/>
       </SwiperSlide>
       )}
       </Swiper>
