+++ src/components/post/Edit.js
@@ -8,12 +8,14 @@ import swal from "sweetalert";
 // 컴포넌트
 import PostHeader from "./PostHeader";
 import Kakaomap from "../kakaomap/Kakaomap";
-import EditImageSlide from "../imageSlide/EditImageSlide";
+import Title from "./Title";
+import TextBox from "./TextBox";
+
 // 라우터
 import { useNavigate, useParams } from "react-router-dom";
+
 // 아이콘
-import logosky from "../../assets/logosky.png";
-import trashwhite from "../../assets/trashwhite.png";
+import PostSectionPerPlace from "./PostSectionPerPlace";
 
 // 카카오맵
 const { kakao } = window;
@@ -51,7 +53,6 @@ const Edit = (props) => {
   const [content, setContent] = useState(editdata && editdata.content); // 콘텐트 텍스트
   const [inputText, setInputText] = useState(""); // 검색창 검색 키워드
   const [select, setSelect] = useState([]); // 선택한 장소 배열에 담아줌
-  const [allImgUrl, setAllImgUrl] = useState([]); // 장소별 기존 이미지 url과 새로운 이미지 url 모음
   const [imgUrl, setImgUrl] = useState([]); // 선택한 장소 이미지 갯수 카운트 배열
   const [focus, setFocus] = useState(); // 선택한 장소 핀 클릭 목록 포커스
   const [selectedRegion, setRegion] = useState(
@@ -61,37 +62,8 @@ const Edit = (props) => {
   const [selectedPrice, setPrice] = useState(
     editdata && editdata.priceCategory
   ); // 비용 선택
-  const [newImgFile, setNewImgFile] = useState([]); // 이미지 모두 파일
+  const [imgs, setImgs] = useState([]); // 이미지 모두 파일
   const [showPlaceModal, setShowPlaceModal] = useState(false); // 지역모달
-  const region = [
-    "서울",
-    "대전",
-    "경기",
-    "세종",
-    "인천",
-    "대구",
-    "강원도",
-    "울산",
-    "충청도",
-    "광주",
-    "전라도",
-    "부산",
-    "경상도",
-    "제주도",
-  ];
-  const theme = ["힐링", "맛집", "애견동반", "액티비티", "호캉스"];
-  const price = [
-    "10만원 이하",
-    "10만원대",
-    "20만원대",
-    "30만원대",
-    "40만원대",
-    "50만원 이상",
-  ];
-
-  const onClickLeftArrow = () => {
-    navigate("/");
-  };
 
   useEffect(() => {
     if (editdata && editdata.place) {
@@ -100,7 +72,7 @@ const Edit = (props) => {
           if (v.place_name !== imgUrl.place_name) {
             imgUrl.push({
               place_name: v.place_name,
-              imgUrl: [],
+              imgUrl: v.imgUrl,
             });
             return imgUrl;
           }
@@ -129,16 +101,6 @@ const Edit = (props) => {
         });
     }
     list(select);
-    if (editdata && editdata.place) {
-      editdata &&
-        editdata.place.map((v, i) => {
-          allImgUrl.push({
-            place_name: v.place_name,
-            imgUrl: v.imgUrl,
-          });
-          return allImgUrl;
-        });
-    }
     window.scrollTo(0, 0);
   }, [dispatch, editdata]);
 
@@ -155,11 +117,6 @@ const Edit = (props) => {
         });
     }
   }, [dispatch, editdata]);
-
-  // 제목 가져오기
-  const onTitleHandler = (e) => {
-    setTitle(e.currentTarget.value);
-  };
 
   // 선택 장소 목록 모달 open / close
   const openPlaceModal = () => {
@@ -184,16 +141,11 @@ const Edit = (props) => {
     setInputText("");
   };
 
-  // 적힌 콘텐트 텍스트 가져오기
-  const onContentHandler = (e) => {
-    setContent(e.target.value);
-  };
-
   // 첨부이미지 파일들 폼데이터로 담기
   const json = JSON.stringify(select);
   const blob = new Blob([json], { type: "application/json" });
   const editFormData = new FormData();
-  newImgFile.forEach((v, i) => {
+  imgs.forEach((v, i) => {
     editFormData.append("imgUrl", v);
   });
   editFormData.append("title", title);
@@ -211,46 +163,6 @@ const Edit = (props) => {
     searchList_wrap.scrollTo(0,0)
   };
 
-  // 장소 선택 취소
-  const onRemovePlace = (place) => {
-    swal({
-      title: "이 장소를 삭제할까요?",
-      icon: "warning",
-      buttons: true,
-      dangerMode: true,
-    }).then((willDelete) => {
-      if (willDelete) {
-        swal("목록에서 삭제되었습니다", {
-          icon: "success",
-        });
-        if (select && select.length !== 0) {
-          setFocus(select && select[0].place_name);
-        }
-        setSelect((pre) => {
-          const newSelect = pre.filter((v, i) => {
-            return place.place_name !== v.place_name;
-          });
-          list(newSelect);
-          return newSelect;
-        });
-        setImgUrl((pre) => {
-          const imgUrlList = pre.filter((v, i) => {
-            return place.place_name !== v.place_name;
-          });
-          return imgUrlList;
-        });
-        setAllImgUrl((pre) => {
-          const imgUrlList = pre.filter((v, i) => {
-            return place.place_name !== v.place_name;
-          });
-          return imgUrlList;
-        });
-      } else {
-        swal("삭제를 취소했습니다");
-      }
-    });
-  };
-
   // 작성 완료 버튼
   const onHandlerEdit = () => {
     if (select.length === 0) {
@@ -263,7 +175,7 @@ const Edit = (props) => {
       swal("비용을 선택해주세요!");
     } else if (title.length === 0) {
       swal("제목을 적어주세요!");
-    } else if (allImgUrl.length === 0 || newImgFile.lenght === 0) {
+    } else if (imgUrl.length === 0 || imgs.lenght === 0) {
       swal("사진을 첨부해주세요!");
     } else if (content.length < 10) {
       swal("내용은 10자 이상 적어주세요!");
@@ -276,7 +188,6 @@ const Edit = (props) => {
       title
     ) {
       swal("수정 완료하시겠습니까?").then((value) => {
-        swal("수정이 완료되었습니다!");
         dispatch(modifyPostDB(editFormData, param));
         navigate("/");
       });
@@ -376,181 +287,28 @@ const Edit = (props) => {
           place={place}
         />
         {/* 제목 */}
-        <div className="writeTitleWrap">
-          <input
-            type="text"
-            onChange={onTitleHandler}
-            defaultValue={editdata && editdata.title}
-            placeholder="코스 이름을 적어주세요"
-          />
-        </div>
-        {/* 검색하고 선택한 장소가 없을 때 */}
-        {select.length === 0 ? (
-          <div className="sectionWrap">
-            {/* 바뀌는 부분 */}
-            <div className="sectionPerPlace">
-              <div className="sectionPerPlaceWrap">
-                {/* 사진업로드 */}
-                <div className="imgUpload">
-                  {/* 사진업로드하는 장소 이름 */}
-                  <div className="imgUploadHeader">
-                    <div className="imgUploadTitle">
-                      <div className="titleTxtWrap">
-                        <img src={logosky} alt="야너갈 로고" />
-                        최상단 검색창에서 장소를 검색해주세요!
-                      </div>
-                      <div className="clickInfo">
-                        ❗여러 장소를 검색하고 선택할 수 있어요!
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            {/* 텍스트 입력 */}
-            <div className="writeTxt">
-              <textarea
-                placeholder="코스에 대한 설명을 입력해주세요"
-                defaultValue={editdata && editdata.content}
-                onChange={onContentHandler}
-              />
-            </div>
-            <button className="writeSubmit" onClick={onHandlerEdit}>
-              수정 완료하기
-            </button>
-          </div>
-        ) : focus && focus.length !== 0 ? (
-          <div className="sectionWrap">
-            {/* 검색해서 장소를 선택했고 핀을 클릭했을 때 */}
-            {/* 바뀌는 부분 */}
-            <div className="sectionPerPlace">
-              {select &&
-                select.map((l, j) => {
-                  return (
-                    <div
-                      className="sectionPerPlaceWrap"
-                      key={j}
-                      style={
-                        focus === l.place_name
-                          ? { display: "block" }
-                          : { display: "none" }
-                      }
-                    >
-                      {/* 사진업로드 */}
-                      <div className="imgUpload">
-                        {/* 사진업로드하는 장소 이름 */}
-                        <div className="imgUploadHeader">
-                          <div
-                            className="imgUploadTitle"
-                            onClick={openPlaceModal}
-                          >
-                            <div className="titleTxtWrap">
-                              <img src={logosky} alt="야너갈 로고" />
-                              {l.place_name}
-                            </div>
-                            <div className="clickInfo">
-                              ❗여러 장소를 검색하고 선택할 수 있어요!
-                            </div>
-                          </div>
-                          <div
-                            className="removePlaceButton"
-                            onClick={() => {
-                              onRemovePlace(l);
-                            }}
-                          >
-                            <img src={trashwhite} alt="장소 삭제 버튼" />이 장소
-                            삭제
-                          </div>
-                        </div>
-                        <EditImageSlide
-                          editdata={editdata}
-                          select={select}
-                          setSelect={setSelect}
-                          imgUrl={imgUrl}
-                          setImgUrl={setImgUrl}
-                          setNewImgFile={setNewImgFile}
-                          newImgFile={newImgFile}
-                          l={l}
-                          j={j}
-                          allImgUrl={allImgUrl}
-                          setAllImgUrl={setAllImgUrl}
-                          focus={focus}
-                        />
-                      </div>
-                    </div>
-                  );
-                })}
-            </div>
-            {/* 텍스트 입력 */}
-            <div className="writeTxt">
-              <textarea
-                placeholder="코스에 대한 설명을 입력해주세요"
-                defaultValue={editdata && editdata.content}
-                onChange={onContentHandler}
-              />
-            </div>
-            <button className="writeSubmit" onClick={onHandlerEdit}>
-              수정 완료하기
-            </button>
-          </div>
-        ) : (
-          <div className="sectionWrap" id="sectionWrap">
-            {/* 검색해서 장소를 선택했지만 핀을 클릭하지 않았을 때 */}
-            {/* 바뀌는 부분 */}
-            <div className="sectionPerPlace">
-              <div className="sectionPerPlaceWrap">
-                {/* 사진업로드 */}
-                <div className="imgUpload">
-                  {/* 사진업로드하는 장소 이름 */}
-                  <div className="imgUploadHeader">
-                    <div className="imgUploadTitle" onClick={openPlaceModal}>
-                      <div className="titleTxtWrap">
-                        <img src={logosky} alt="야너갈 로고" />
-                        {select && select[0] && select[0].place_name}
-                      </div>
-                      <div className="clickInfo">
-                        ❗여러 장소를 검색하고 선택할 수 있어요!
-                      </div>
-                    </div>
-                    <div
-                      className="removePlaceButton"
-                      onClick={() => {
-                        onRemovePlace(select && select[0]);
-                      }}
-                    >
-                      <img src={trashwhite} alt="장소 삭제 버튼" />이 장소 삭제
-                    </div>
-                  </div>
-                  <EditImageSlide
-                    editdata={editdata}
-                    select={select}
-                    setSelect={setSelect}
-                    imgUrl={imgUrl}
-                    setImgUrl={setImgUrl}
-                    setNewImgFile={setNewImgFile}
-                    newImgFile={newImgFile}
-                    l={select && select[0]}
-                    j={0}
-                    allImgUrl={allImgUrl}
-                    setAllImgUrl={setAllImgUrl}
-                    focus={focus}
-                  />
-                </div>
-              </div>
-            </div>
-            {/* 텍스트 입력 */}
-            <div className="writeTxt">
-              <textarea
-                placeholder="코스에 대한 설명을 입력해주세요"
-                defaultValue={editdata && editdata.content}
-                onChange={onContentHandler}
-              />
-            </div>
-            <button className="writeSubmit" onClick={onHandlerEdit}>
-              수정 완료하기
-            </button>
-          </div>
-        )}
+        <Title setTitle={setTitle} param={param} editdata={editdata}/>
+
+        {/* 장소마다 다른 부분 */}
+        <PostSectionPerPlace
+          select={select}
+          setSelect={setSelect}
+          focus={focus}
+          setFocus={setFocus}
+          list={list}
+          imgUrl={imgUrl}
+          setImgUrl={setImgUrl}
+          openPlaceModal={openPlaceModal}
+          setImgs={setImgs}
+          imgs={imgs}
+          editdata={editdata}
+        />
+
+        {/* 텍스트 입력 */}
+        <TextBox editdata={editdata} setContent={setContent} param={param}/>
+        <button className="writeSubmit" onClick={onHandlerEdit}>
+          수정 완료하기
+        </button>
       </div>
     </div>
   );
