+++ src/css/post.scss
@@ -287,6 +287,7 @@
 
     .sectionPerPlace{
       width: 343px;
+      margin-bottom: 18px;
 
       .sectionPerPlaceWrap{
         display: flex;
